+++ DesignPatterns/SimpleFactoryDemo/Models/Student.cs
@@ -5,7 +5,7 @@
 
     public abstract class Student : IStudent
     {
-        protected Student(string name, IList<int> evaluations)
+        protected Student(string name, IList<double> evaluations)
         {
             this.Name = name;
             this.Evaluations = evaluations;
@@ -13,6 +13,6 @@
 
         public string Name { get; private set; }
 
-        public IList<int> Evaluations { get; private set; }
+        public IList<double> Evaluations { get; private set; }
     }
 }
